Open SD folder with Thonny
I snipped out this section of code from the main program for convenient use in Thonny. When the Main program crashes, but you need to see what was written to the SD card, running this program makes the SD card contents available in Thonny. Simply use the File Open command and browse to the sd folder on the Pico. Opening the folder may take some time if you have many files.